requirements.txt
numpy==1.16.4
google-cloud-automl==0.4.0
google-cloud-storage==1.17.0
google-api-core==1.14.2
google-auth==1.6.3
seaborn==0.9.0
pandas==0.24.2
matplotlib==3.1.1
jupyter==1.0.0
prompt-toolkit
setuptools==41.0.1
pytest==5.0.1
pytest-remotedata==0.3.2
numpy==1.16.4
pyts==0.8.0
numba==0.45.0
pyspark==2.4.3
findspark==1.3.0
PyArrow==0.14.0
redis==3.3.5
stopwatch.py==1.0.1
absl-py==0.7.0
astor==0.6
gast==0.2.0
google-pasta==0.1.6
grpcio==1.22.0
keras-applications==1.0.6
keras-preprocessing==1.0.5
protobuf==3.6.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
msgpack==0.6.1
wrapt==1.11.1
Pillow==6.1.0
tensorflow-gpu==1.14.0
pywin32==224
wget==3.2
quandl==3.4.8
more-itertools==5.0.0
requests==2.22.0
tqdm==4.34.0
# NOTE: 2019-08-26; More recent
scikit-learn==0.15.0